Added a bunch of type checking and fixed some coastal issues

+++ DiploGM/parse_edit_state.py
@@ -44,7 +44,7 @@ logger = logging.getLogger(__name__)
 manager = Manager()
 
 
-def parse_edit_state(message: str, board: Board) -> dict[str, str | bytes | int | None]:
+def parse_edit_state(message: str, board: Board) -> tuple[str, str, bytes | None, str | None, str | None]:
     invalid: list[tuple[str, Exception]] = []
     commands = str.splitlines(message)
     for command in commands:
@@ -73,13 +73,13 @@ def parse_edit_state(message: str, board: Board) -> dict[str, str | bytes | int 
     else:
         file, file_name = None, None
 
-    return {
-        "title": response_title,
-        "message": response_body,
-        "file": file,
-        "file_name": file_name,
-        "embed_colour": embed_colour,
-    }
+    return (
+        response_title,
+        response_body,
+        file,
+        file_name,
+        embed_colour,
+    )
 
 
 def _parse_command(command: str, board: Board) -> None:
@@ -244,9 +244,10 @@ def _create_unit(keywords: list[str], board: Board) -> None:
     if not player:
         raise ValueError(f"Unknown player: {keywords[1]}")
     province, coast = board.get_province_and_coast(" ".join(keywords[2:]))
-    # if unit_type == UnitType.FLEET and coast is None:
-    #     coast_name = f"{province} coast"
-    #     province, coast = board.get_province_and_coast(coast_name)
+    if province.get_multiple_coasts() and coast not in province.get_multiple_coasts():
+        raise ValueError(f"Province '{province.name}' requires a valid coast.")
+    if not province.get_multiple_coasts():
+        coast = None
 
     unit = board.create_unit(unit_type, player, province, coast, None)
     get_connection().execute_arbitrary_sql(
@@ -275,6 +276,10 @@ def _create_dislodged_unit(keywords: list[str], board: Board) -> None:
         if not player:
             raise ValueError(f"Unknown player: {keywords[1]}")
         province, coast = board.get_province_and_coast(keywords[2])
+        if province.get_multiple_coasts() and coast not in province.get_multiple_coasts():
+            raise ValueError(f"Province '{province.name}' requires a valid coast.")
+        if not province.get_multiple_coasts():
+            coast = None
         retreat_options = set(
             [board.get_province_and_coast(province_name) for province_name in keywords[3:]]
         )
@@ -351,6 +356,10 @@ def _move_unit(keywords: list[str], board: Board) -> None:
     if not unit:
         raise RuntimeError(f"No unit to move in {old_province}")
     new_province, new_coast = board.get_province_and_coast(keywords[1])
+    if new_province.get_multiple_coasts() and new_coast not in new_province.get_multiple_coasts():
+        raise ValueError(f"Province '{new_province.name}' requires a valid coast.")
+    if not new_province.get_multiple_coasts():
+        new_coast = None
     board.move_unit(unit, new_province, new_coast)
     get_connection().execute_arbitrary_sql(
         "DELETE FROM units WHERE board_id=? and phase=? and location=? and is_dislodged=?",
